+++ src/options.rs
@@ -44,6 +44,9 @@ pub struct Options {
     /// taken by the baseline test.
     pub test_timeout: Option<Duration>,
 
+    /// The time multiplier for test tasks, if set (relative to baseline test duration).
+    pub test_timeout_multiplier: Option<f64>,
+
     /// The minimum test timeout, as a floor on the autoset value.
     pub minimum_test_timeout: Duration,
 
@@ -211,6 +214,7 @@ impl Options {
             show_times: !args.no_times,
             show_all_logs: args.all_logs,
             test_timeout: args.timeout.map(Duration::from_secs_f64),
+            test_timeout_multiplier: config.timeout_multiplier.or(args.timeout_multiplier),
             test_tool: args.test_tool.or(config.test_tool).unwrap_or_default(),
         };
         options.error_values.iter().for_each(|e| {